pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "heuristix"
version = "0.1.0"
description = "Knowledge-Evolutionary Framework for Automatic Heuristic Discovery"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27",
    "pyyaml>=6.0",
    "numpy>=1.24",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.4"]
viz = ["matplotlib>=3.8"]

[project.scripts]
heuristix = "heuristix.cli:main"
